refactor: rename filename for quote interface
