+++ src/lesson/lesson.service.ts
@@ -12,7 +12,7 @@ import { AlsStore } from '#common/als/store-validator.js';
 import AuthExceptionMessage from '#exception/auth-exception-message.js';
 import LessonExceptionMessage from '#exception/lesson-exception-message.js';
 import { UserService } from '#user/user.service.js';
-import type { IQuoteService } from '#quote/interface/quote-service.inteface.js';
+import type { IQuoteService } from '#quote/interface/quote-service.interface.js';
 import { CreateDirectQuoteDto, QueryLessonDto, RejectDirectQuoteDto } from './dto/lesson.dto.js';
 import type { ILessonService } from './interface/lesson-service.interface.js';
 import { LessonRepository } from './lesson.repository.js';
