+++ src/quote/interface/quote-service.interface.ts
